MODIFICATION : my x and X || ADD : my p

+++ ft_hexa_nbr.c
@@ -6,17 +6,17 @@
 /*   By: vicperri <vicperri@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/18 11:45:40 by vicperri          #+#    #+#             */
-/*   Updated: 2024/11/18 17:11:30 by vicperri         ###   ########lyon.fr   */
+/*   Updated: 2024/11/19 16:18:11 by vicperri         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_hexa_nbr(unsigned long n, char x)
+int	ft_hexa_nbr(unsigned int n, char x)
 {
-	char *base_maj;
-	char *base_min;
-	int count;
+	char	*base_maj;
+	char	*base_min;
+	int		count;
 
 	count = 0;
 	base_maj = "0123456789ABCDEF";
